+++ src/game/items/container.ts
@@ -33,7 +33,8 @@ export function newContainer(config: ContainerConfig & ItemConfig) {
     size,
     closeable,
     lockable,
-    key
+    key,
+    config
   );
 
   if (verbs) {
@@ -78,7 +79,8 @@ export class Container extends Item {
     size = 1,
     closeable = true,
     lockable = false,
-    key?: string | KeyT
+    key?: string | KeyT,
+    config?: ContainerConfig
   ) {
     const dynamicOpenDescription = createDynamicText(openDescription);
     const dynamicClosedDescription = createDynamicText(closedDescription);
@@ -88,7 +90,9 @@ export class Container extends Item {
       holdable,
       size,
       [],
-      aliases || []
+      aliases || [],
+      undefined,
+      config
     );
     this.canHoldItems = true;
     this.capacity = capacity;
@@ -99,14 +103,14 @@ export class Container extends Item {
     this.closeable = closeable;
     this.lockable = lockable;
     this.key = key;
-    this.lockedText = `The ${this.name} is locked.`;
+    this.lockedText = `The ${this.name} ${this.isOrAre()} locked.`;
     this.openText = `The ${this.name} opens easily.`;
-    this.alreadyOpenText = `The ${this.name} is already open.`;
+    this.alreadyOpenText = `The ${this.name} ${this.isOrAre()} already open.`;
     this.closeText = `You close the ${this.name} with a soft thud.`;
-    this.alreadyClosedText = `The ${this.name} is already closed.`;
+    this.alreadyClosedText = `The ${this.name} ${this.isOrAre()} already closed.`;
     this.wrongKeyText = `The key doesn't fit.`;
     this.needsKeyText = `The ${name} appears to need a key.`;
-    this.alreadyUnlockedText = `The ${name} is already unlocked.`;
+    this.alreadyUnlockedText = `The ${name} ${this.isOrAre()} already unlocked.`;
 
     if (this.closeable) {
       this.openVerb = new Verb.Builder("open")