+++ changelog.md
@@ -1,13 +1,17 @@
 # goNORTH Changelog
 
+# Unreleased
+
+* 2024-08-04 Items can be made plural and 'isOrAre' function added.
+
 ## 3.5.0
 
-* 2024-07-04 Can override default peek behaviour, doors can be transparent.
-* 2024-07-03 Ability to peek through doors.
+* 2024-08-04 Can override default peek behaviour, doors can be transparent.
+* 2024-08-03 Ability to peek through doors.
 
 # 3.4.0
 
-* 2024-07-02 Items that produce singular items trigger a 'take' action and then defer verbs to their children.
+* 2024-08-02 Items that produce singular items trigger a 'take' action and then defer verbs to their children.
 * 2024-07-31 Used correct item article in plural item smart test message.
 
 ## 3.3.0